# Add nearest shape to point queries for bvh