--- a/src/aabb/aabb_impl.rs
+++ b/src/aabb/aabb_impl.rs
@@ -594,6 +594,53 @@ impl<T: BHValue, const D: usize> Aabb<T, D> {
     pub fn largest_axis(&self) -> usize {
         self.size().imax()
     }
+
+    /// Returns the minimum and maximum distances to the [`Aabb`].
+    /// The minimum distance is the distance to the closest point on the box,
+    /// 0 if the point is inside the box.
+    /// The maximum distance is the distance to the furthest point in the box.
+    ///
+    /// # Examples
+    /// ```
+    /// use bvh::aabb::Aabb;
+    /// use nalgebra::Point3;
+    ///
+    /// let min = Point3::new(0.0,0.0,0.0);
+    /// let max = Point3::new(10.0,10.0,10.0);
+    ///
+    /// let aabb = Aabb::with_bounds(min, max);
+    /// let query = Point3::new(20.0, 0.0, 0.0);
+    /// let (min_dist, max_dist) = aabb.get_min_max_distances(&query);
+    /// assert!(min_dist == 10.0);
+    /// assert!(max_dist == (10.0_f32*10.0 + 10.0*10.0 + 20.0*20.0).sqrt());
+    /// ```
+    ///
+    /// [`Aabb`]: struct.Aabb.html
+    ///
+    pub fn get_min_max_distances(&self, point: &Point<T, D>) -> (T, T) {
+        let half_size = self.size() * T::from_f32(0.5).unwrap();
+        let center = self.center();
+
+        let delta = point - center;
+
+        // See <https://iquilezles.org/articles/distfunctions/>
+        let q = delta.abs() - half_size;
+        let outside_vec = q.map(|x| x.max(T::zero()));
+        let min_dist = outside_vec.dot(&outside_vec).sqrt(); // norm without requiring ComplexField.
+
+        // The signum helps to determine the furthest point
+        let signum = delta
+            // Invert the signum to get the furthest vertex
+            .map(|x| -x.signum())
+            // Make sure we're always on a vertex and not on a face if the point is aligned with the box
+            .map(|x| if x != T::zero() { x } else { T::one() });
+
+        let furthest = center + signum.component_mul(&half_size);
+        let furthest_delta = point - furthest;
+        let max_dist = furthest_delta.dot(&furthest_delta).sqrt();
+
+        (min_dist, max_dist)
+    }
 }
 
 /// Default instance for [`Aabb`]s. Returns an [`Aabb`] which is [`empty()`].