I recently implemented a distance point query with a bvh for my crate mesh_to_sdf by trait extending this crate (see https://github.com/Azkellas/mesh_to_sdf/pull/75). This PR is an attempt to upstream the changes and would solve https://github.com/svenstaro/bvh/issues/56.

This PR is different from https://github.com/svenstaro/bvh/pull/43 which only returns shapes with aabb containing the query point. On the contrary, this PR returns a conservative list of candidates, one of which for sure contains the shape that is closest to the query point. This is needed since we can't know for sure if the query point is in the aabb of its closest object.

To do so, we keep track of the best minimum/maximum distance found yet in the tree. For each node we compute the aabb min/max distance from the point (the minimum is 0 if the point is inside). Then we can prune or investigate the subtrees depending on the aabb's min/max compared to the current best min/max.

This means the search visits a subtree and not a single linear branch. It might be interesting to add an `approx` version following https://github.com/svenstaro/bvh/pull/43 implementation which would be faster and "close enough".

I couldn't implement the `nearest_candidate` method for `FlatBvh`.  Pre-order no-stack no-recursion tree traversal requires to store the parent index in the node afaik. I didn't want to change the memory layout without talking about it first. This PR is in draft state for this reason.

Testing is a bit rudimentary (and no benchmark) as I did most of it in my crate and haven't ported it yet.
Anyway for 100k triangles and 15k queries (in mesh's bounding box) it is 10x faster than brute forcing while being exact.

Waht remains to be done:
- [x] Upstream tests and benches
- [x] Implement `nearest_candidate` for `FlatBvh`
- [x] Implement an `approx` version? At least test the performance gain.

Thank you for the work, this crate is very useful.