@svenstaro 2024-07-04
Thanks a lot! I will take an in-depth look soon. How do you think this interacts with #98? I think I would like to merge #98 first and I'm almost done with it but I don't want to negate your work. Could you take a look at let me know what you think?