@Azkellas 2024-07-05
I think the two PR are mostly orthogonal and merging #98 shouldn't affect this PR much.
Interestingly, they share some similarities in the tree traversal, but I think it's better to keep them separate as the `nearest_candidates` does not need ordering and thus gives more room for optimization. My point-bvh distance query also suffers from the fact that it needs to visit aabbs that don't contain the query point. 
This PR also doesn't provide an iterator since it uses the whole result to prune the vector (often removing ~75% of the candidates).