@svenstaro 2024-07-05
In a related note: Would you like to help me maintain this library? It's not getting enough love and could do with some better maintenance.