@svenstaro 2024-07-06
I merged and released #98 so this one now needs a rebase. :)