@Azkellas 2024-07-07
I rebased without issue.
This is what the benches look like at the moment.

```
test bvh::bvh_impl::bench::bench_nearest_candidates_1200_triangles_bvh        ... bench:       1,674.28 ns/iter (+/- 82.26)
test bvh::bvh_impl::bench::bench_nearest_candidates_120k_triangles_bvh        ... bench:      15,676.73 ns/iter (+/- 838.18)
test bvh::bvh_impl::bench::bench_nearest_candidates_12k_triangles_bvh         ... bench:       4,685.55 ns/iter (+/- 165.48)
test bvh::bvh_impl::bench::bench_nearest_candidates_sponza_bvh                ... bench:      47,883.61 ns/iter (+/- 4,195.02)
test testbase::bench_nearest_candidates_120k_triangles_list                   ... bench:     664,322.50 ns/iter (+/- 39,013.88)
test testbase::bench_nearest_candidates_120k_triangles_list_aabb              ... bench:     519,004.38 ns/iter (+/- 63,937.19)
test testbase::bench_nearest_candidates_sponza_list                           ... bench:     611,325.62 ns/iter (+/- 33,850.31)
test testbase::bench_nearest_candidates_sponza_list_aabb                      ... bench:     282,300.00 ns/iter (+/- 41,833.67)
```

I'm not against becoming a co-maintainer, we can talk about it to see what needs to be done.