@Azkellas 2024-07-09
I implemented the function for flat bvh as well. The performances are equivalent. I think the PR is ready for review now. Not perfect but functional.

> [ ] Upstream tests and benches
> [x] Implement nearest_candidate for FlatBvh
> [ ] Implement an approx version? At least test the performance gain.

Regarding the checklist in the first post, benches are here, tests are still a bit lackluster.
The approx version actually doesn't make sense since for a query point outside of the bvh bounding box it would return an empty list.