@Azkellas 2025-01-07
Nice to see some activity on the project! I have now a faster implementation on my crate (though not public yet) that simplifies the algorithm by asking the user for more information.

This PR finds nearest candidates by only using aabbs (requiring no more information from the user), while my latest implementation add a trait to query the exact distance on relevant leaves. This allows for a way better pruning and, IIRC, makes the whole query up to 3x faster for triangles. 

I'll try to make/update a PR this week.