@finnbear 2025-01-07
Hi :)

> I'll try to make/update a PR this week.

Does it still make sense to review & merge this PR as-is?