@Azkellas 2025-01-07
The two versions could cohabit (nearest candidates could be useful is the cost of computing the exact distance is really high), otherwise the new version should substitute this one.
Though if the cost is high it probably means the new version would prune better too.

I think you can can wait for my update before you try to rebase this version. :)