@Azkellas 2025-01-07
> Sorry again for the long wait, life got busy. :( 

No problem, I turned on my computed for the first time in three months last Wednesday, so we're quite synchronized at least. :) 

> So as I understand it, you will create a new PR? Shall we close this one then? 

I'll see if it's easier to rebase this one or upstream from scratch.