@Azkellas 2025-01-12
I updated the code with the latest implementation.

Now instead of keeping a limit (min, max) on the distance to the best shape, we keep track of the actual distance to the current best shape (taking the first shape arbitrarily  if the bvh is not empty).

The two good news are:
- The implementation is much easier to read and maintain than before
- Performance is much much better, from 6x faster for the 1200 triangles benchmark to 50x faster for sponza (the more triangles, the faster), and 1000x faster than the most naive implementation.

The two bad news are:
- This adds the `PointDistance` trait requirement for Shape, needed to query the exact distance (squared) from the shape to the query point on candidate leaf nodes.
- The flat bvh is not optimized for distance queries as the function cannot choose which child to check first. The flat bvh is ~5-10x slower depending on the size of the bvh (still 100x faster than a naive implementation).

I think the branch is ready for review.

<details>
<summary>Incremental benchmarks</summary>

```
// naive implementation: go through each shape one by one with no BVH
test testbase::bench_nearest_candidates_120k_triangles_list                   ... bench:     681,288.75 ns/iter (+/- 55,478.50)
test testbase::bench_nearest_candidates_120k_triangles_list_aabb              ... bench:     534,145.00 ns/iter (+/- 98,617.50)
test testbase::bench_nearest_candidates_sponza_list                           ... bench:     619,938.12 ns/iter (+/- 39,642.44)
test testbase::bench_nearest_candidates_sponza_list_aabb                      ... bench:     295,907.50 ns/iter (+/- 39,118.38)
```
```
// old version with aabb min/max distances only
test bvh::bvh_impl::bench::bench_nearest_candidates_1200_triangles_bvh        ... bench:       1,744.73 ns/iter (+/- 80.47)
test bvh::bvh_impl::bench::bench_nearest_candidates_120k_triangles_bvh        ... bench:      15,921.70 ns/iter (+/- 1,739.16)
test bvh::bvh_impl::bench::bench_nearest_candidates_12k_triangles_bvh         ... bench:       4,841.70 ns/iter (+/- 399.85)
test bvh::bvh_impl::bench::bench_nearest_candidates_sponza_bvh                ... bench:      50,180.16 ns/iter (+/- 6,593.96)
```
```
// first version with exact distance query on candidate leaf nodes
test bvh::bvh_impl::bench::bench_nearest_from_1200_triangles_bvh              ... bench:       1,539.03 ns/iter (+/- 116.92)
test bvh::bvh_impl::bench::bench_nearest_from_120k_triangles_bvh              ... bench:      15,680.75 ns/iter (+/- 1,989.23)
test bvh::bvh_impl::bench::bench_nearest_from_12k_triangles_bvh               ... bench:       4,207.26 ns/iter (+/- 454.26)
test bvh::bvh_impl::bench::bench_nearest_from_sponza_bvh                      ... bench:       8,367.80 ns/iter (+/- 640.04)
```
```
// take first shape as first guess instead of T::max
test bvh::bvh_impl::bench::bench_nearest_from_1200_triangles_bvh              ... bench:       1,101.38 ns/iter (+/- 89.26)
test bvh::bvh_impl::bench::bench_nearest_from_120k_triangles_bvh              ... bench:      11,637.48 ns/iter (+/- 1,637.42)
test bvh::bvh_impl::bench::bench_nearest_from_12k_triangles_bvh               ... bench:       3,053.51 ns/iter (+/- 407.65)
test bvh::bvh_impl::bench::bench_nearest_from_sponza_bvh                      ... bench:       5,031.79 ns/iter (+/- 456.78)
```
```
// go to best min child, useful to get a better first estimation of the best child.
test bvh::bvh_impl::bench::bench_nearest_from_1200_triangles_bvh              ... bench:         276.45 ns/iter (+/- 17.75)
test bvh::bvh_impl::bench::bench_nearest_from_120k_triangles_bvh              ... bench:         776.33 ns/iter (+/- 112.84)
test bvh::bvh_impl::bench::bench_nearest_from_12k_triangles_bvh               ... bench:         426.19 ns/iter (+/- 20.88)
test bvh::bvh_impl::bench::bench_nearest_from_sponza_bvh                      ... bench:       1,211.11 ns/iter (+/- 64.11)
```

```
// comparison with flat bvh
test bvh::bvh_impl::bench::bench_nearest_from_1200_triangles_bvh              ... bench:         284.03 ns/iter (+/- 21.04)
test bvh::bvh_impl::bench::bench_nearest_from_120k_triangles_bvh              ... bench:         837.90 ns/iter (+/- 264.71)
test bvh::bvh_impl::bench::bench_nearest_from_12k_triangles_bvh               ... bench:         427.79 ns/iter (+/- 23.38)
test flat_bvh::bench::bench_nearest_from_1200_triangles_bvh                   ... bench:         700.26 ns/iter (+/- 34.75)
test flat_bvh::bench::bench_nearest_from_120k_triangles_bvh                   ... bench:       8,509.34 ns/iter (+/- 1,119.36)
test flat_bvh::bench::bench_nearest_from_12k_triangles_bvh                    ... bench:       2,148.05 ns/iter (+/- 179.71)
```
</details>